--- src/main/java/org/example/o2o/api/v1/controller/menu/MenuOptionController.java
@@ -1,48 +1,0 @@
-package org.example.o2o.api.v1.controller.menu;
-
-import org.example.o2o.api.v1.docs.menu.MenuOptionDocsController;
-import org.example.o2o.api.v1.dto.menu.request.MenuOptionGroupCreateRequestDto;
-import org.example.o2o.api.v1.dto.menu.response.MenuOptionGroupResponseDto;
-import org.example.o2o.api.v1.service.menu.MenuOptionService;
-import org.example.o2o.common.dto.ApiResponse;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@RestController
-@RequestMapping("/api/v1")
-public class MenuOptionController implements MenuOptionDocsController {
-
-	private final MenuOptionService menuOptionService;
-
-	@PostMapping("/menus/{menuId}/options")
-	public ApiResponse<MenuOptionGroupResponseDto> registerMenuOption(@PathVariable(name = "menuId") Long menuId,
-		@RequestBody MenuOptionGroupCreateRequestDto requestDto) {
-
-		return ApiResponse.success(menuOptionService.register(menuId, requestDto.toStoreMenuOptionGroup()));
-	}
-
-	@DeleteMapping("/options/{optionGroupId}")
-	public ApiResponse<Void> deleteMenuOption(
-		@PathVariable(name = "optionGroupId") Long optionGroupId) {
-
-		menuOptionService.delete(optionGroupId);
-
-		return ApiResponse.success();
-	}
-
-	@PatchMapping("/options/{optionGroupId}")
-	public ApiResponse<MenuOptionGroupResponseDto> updateMenuOption(
-		@PathVariable(name = "optionGroupId") Long optionGroupId,
-		@RequestBody MenuOptionGroupCreateRequestDto requestDto) {
-
-		return ApiResponse.success(menuOptionService.update(optionGroupId, requestDto.toStoreMenuOptionGroup()));
-	}
-}
