+++ src/main/java/org/example/o2o/api/v2/dto/menu/response/MenuOptionGroupResponseDto.java
@@ -1,10 +1,9 @@
-package org.example.o2o.api.v1.dto.store.response;
+package org.example.o2o.api.v2.dto.menu.response;
 
 import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import org.example.o2o.api.v2.dto.menu.response.MenuOptionResponseDto;
 import org.example.o2o.domain.menu.StoreMenuOption;
 import org.example.o2o.domain.menu.StoreMenuOptionGroup;
 
@@ -12,9 +11,12 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 
 @Builder
-public record StoreMenuOptionGroupResponse(
+public record MenuOptionGroupResponseDto(
 	@Schema(description = "메뉴 옵션 ID", example = "1")
 	long optionGroupId,
+
+	@Schema(description = "메뉴 옵션 정렬 순서", example = "1")
+	int ordering,
 
 	@Schema(description = "메뉴 옵션 필수 여부", example = "true")
 	boolean isRequired,
@@ -23,18 +25,18 @@ public record StoreMenuOptionGroupResponse(
 	boolean isMultiple,
 
 	@Schema(description = "메뉴 옵션명", example = "맵기 조절")
-	String optionGroupName,
+	String title,
 
 	@Schema(description = "메뉴 옵션 항목 목록", example = "[]")
 	List<MenuOptionResponseDto> options
 ) {
-	public static StoreMenuOptionGroupResponse of(
-		StoreMenuOptionGroup optionGroup) {
-		return StoreMenuOptionGroupResponse.builder()
+	public static MenuOptionGroupResponseDto of(StoreMenuOptionGroup optionGroup) {
+		return MenuOptionGroupResponseDto.builder()
 			.optionGroupId(optionGroup.getId())
+			.ordering(optionGroup.getOrdering())
 			.isRequired(optionGroup.getIsRequired())
 			.isMultiple(optionGroup.getIsMultiple())
-			.optionGroupName(optionGroup.getTitle())
+			.title(optionGroup.getTitle())
 			.options(optionGroup.getOptions()
 				.stream()
 				.sorted(Comparator.comparingInt(StoreMenuOption::getPrice)) // TODO ordering으로 정렬
